pyproject.toml
[project]
name = "recruitment-tool"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aiohttp==3.9.0",
    "aiomysql==0.2.0",
    "aiosignal==1.3.1",
    "async-timeout==4.0.2",
    "attrs==22.2.0",
    "autopep8==2.0.2",
    "beautifulsoup4==4.11.2",
    "bs4==0.0.1",
    "certifi==2022.12.7",
    "cffi==1.17.1",
    "charset-normalizer==3.1.0",
    "cron-converter==1.0.2",
    "cryptography==44.0.0",
    "dataclasses-json==0.5.7",
    "discord-py==2.3.2",
    "frozenlist==1.4.1",
    "idna==3.4",
    "lxml==4.9.3",
    "marshmallow==3.19.0",
    "marshmallow-enum==1.5.1",
    "multidict==6.1.0",
    "mypy-extensions==1.0.0",
    "packaging==23.0",
    "pycodestyle==2.10.0",
    "pycparser==2.22",
    "pymysql==1.1.0",
    "python-dateutil==2.8.2",
    "python-dotenv==1.0.0",
    "requests==2.28.2",
    "six==1.16.0",
    "soupsieve==2.4",
    "typing-extensions==4.5.0",
    "typing-inspect==0.8.0",
    "urllib3==1.26.15",
    "yarl==1.9.3",
]
